+++ app/src/main/java/com/nike/snkrs/sunderland/ui/view/AdapterLocalData.kt
@@ -11,14 +11,13 @@ package com.nike.snkrs.sunderland.ui.view
 //region import directives
 
 import android.animation.Animator
-import android.net.Uri
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.nike.snkrs.sunderland.databinding.CardDataBinding
+import com.nike.snkrs.sunderland.databinding.CardLocalDataBinding
 import com.nike.snkrs.sunderland.ui.view.callbacks.AdapterDataCallback
 import com.nike.snkrs.sunderland.util.GlideApp
 import com.nike.snkrs.sunderland.util.tryCatch
@@ -28,11 +27,11 @@ import com.nike.snkrs.sunderland.util.tryCatchWithLogging
 
 
 /**
- * General data adapter for presenting cards with images & text
+ * Data adapter for presenting local data on cards with images (primary and alternates) & text (source)
  * @author Thomas Sunderland. 2021 MAY 12
  */
-class AdapterData(val callback: AdapterDataCallback) :
-    ListAdapter<AdapterDataItem, AdapterData.ViewHolder>(AdapterDataItemDiffCallBack()) {
+class AdapterLocalData(val callback: AdapterDataCallback) :
+    ListAdapter<AdapterLocalDataItem, AdapterLocalData.ViewHolder>(AdapterDataItemDiffCallBack()) {
 
     //region inner types
 
@@ -40,19 +39,15 @@ class AdapterData(val callback: AdapterDataCallback) :
      * ViewHolder
      */
     //@formatter:off
-    inner class ViewHolder(val binding: CardDataBinding,
+    inner class ViewHolder(val binding: CardLocalDataBinding,
                            private val callback: AdapterDataCallback) : RecyclerView.ViewHolder(binding.root) {
 
-        fun bind(item: AdapterDataItem) {
+        fun bind(item: AdapterLocalDataItem) {
             with(binding) {
                 tryCatch {
-                    // load image
-                    // if the resource is an int then we know it's a local drawable resource id,
-                    // otherwise we'll try parsing it as a URI (remote image)
+                    // load primary image
                     item.resource.toIntOrNull()?.let {
                         GlideApp.with(root).load(it).into(featuredImage)
-                    } ?: run {
-                        GlideApp.with(root).load(Uri.parse(item.resource)).into(featuredImage)
                     }
 
                     // set image source
@@ -90,7 +85,7 @@ class AdapterData(val callback: AdapterDataCallback) :
      * Implementation of DiffUtil.ItemCallback for ensuring that items are only updated as-needed
      */
     //@formatter:off
-    private class AdapterDataItemDiffCallBack : DiffUtil.ItemCallback<AdapterDataItem>() {
+    private class AdapterDataItemDiffCallBack : DiffUtil.ItemCallback<AdapterLocalDataItem>() {
         /**
          * Called to check whether two objects represent the same item.
          *
@@ -98,7 +93,7 @@ class AdapterData(val callback: AdapterDataCallback) :
          * @param newItem The item in the new list.
          * @return True if the two items represent the same object or false if they are different.
          */
-        override fun areItemsTheSame(oldItem: AdapterDataItem, newItem: AdapterDataItem): Boolean =
+        override fun areItemsTheSame(oldItem: AdapterLocalDataItem, newItem: AdapterLocalDataItem): Boolean =
             oldItem === newItem || oldItem.id == newItem.id
 
         /**
@@ -108,7 +103,7 @@ class AdapterData(val callback: AdapterDataCallback) :
          * @param newItem The item in the new list.
          * @return True if the contents of the items are the same or false if they are different.
          */
-        override fun areContentsTheSame(oldItem: AdapterDataItem, newItem: AdapterDataItem): Boolean =
+        override fun areContentsTheSame(oldItem: AdapterLocalDataItem, newItem: AdapterLocalDataItem): Boolean =
             oldItem == newItem
     }
     //@formatter:on
@@ -123,8 +118,9 @@ class AdapterData(val callback: AdapterDataCallback) :
      * @param viewType The view type of the new View.
      * @return A new ViewHolder that holds a View of the given view type.
      */
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AdapterData.ViewHolder {
-        val binding = CardDataBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AdapterLocalData.ViewHolder {
+        val binding =
+            CardLocalDataBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding, callback)
     }
 
@@ -133,7 +129,7 @@ class AdapterData(val callback: AdapterDataCallback) :
      * @param viewHolder The ViewHolder which should be updated to represent the contents of the item at the given position in the data set.
      * @param position The position of the item within the adapter's data set.
      */
-    override fun onBindViewHolder(viewHolder: AdapterData.ViewHolder, position: Int) {
+    override fun onBindViewHolder(viewHolder: AdapterLocalData.ViewHolder, position: Int) {
         tryCatch { viewHolder.bind(getItem(position % currentList.size)) }
     }